storage_service/requirements.txt
cassandra-driver==3.25.0
psycopg2==2.9.1
